+++ packages/playground/blueprints/public/blueprint-schema.json
@@ -190,6 +190,9 @@
 						},
 						"disableWpNewBlogNotification": {
 							"type": "boolean"
+						},
+						"makeEditorFrameControlled": {
+							"type": "boolean"
 						}
 					},
 					"required": ["step"]
