+++ packages/playground/compile-wordpress/Dockerfile
@@ -43,43 +43,6 @@ RUN cp -r wordpress wordpress-static && \
     # Remove all empty directories
     find . -type d -empty -delete
 
-# Pair the site editor's nested iframe to the Service Worker.
-# 
-# Without the patch below, the site editor initiates network requests that
-# aren't routed through the service worker. That's a known browser issue:
-# 
-# * https://bugs.chromium.org/p/chromium/issues/detail?id=880768
-# * https://bugzilla.mozilla.org/show_bug.cgi?id=1293277
-# * https://github.com/w3c/ServiceWorker/issues/765
-# 
-# The problem with iframes using srcDoc and src="about:blank" as they
-# fail to inherit the root site's service worker.
-# 
-# Gutenberg loads the site editor using <iframe srcDoc="<!doctype html">
-# to force the standards mode and not the quirks mode:
-# 
-# https://github.com/WordPress/gutenberg/pull/38855
-# 
-# This commit patches the site editor to achieve the same result via
-# <iframe src="/doctype.html"> and a doctype.html file containing just
-# `<!doctype html>`. This allows the iframe to inherit the service worker
-# and correctly load all the css, js, fonts, images, and other assets.
-# 
-# Ideally this issue would be fixed directly in Gutenberg and the patch
-# below would be removed.
-# 
-# See https://github.com/WordPress/wordpress-playground/issues/42 for more details
-COPY ./build-assets/controlled-iframe.js /root/
-
-RUN echo '<!doctype html><script>const hash = window.location.hash.substring(1); if ( hash ) document.write(decodeURIComponent(hash))</script>' > wordpress-static/wp-includes/empty.html
-# Replace a vanilla HTML iframe with Playground's ControlledIframe component in the block editor.
-RUN sed -E 's#\(\s*"iframe",#(__playground_ControlledIframe,#g' -i wordpress-static/wp-includes/js/dist/block-editor.js && \
-    cat /root/controlled-iframe.js wordpress-static/wp-includes/js/dist/block-editor.js >> /root/block-editor.js && \
-    mv /root/block-editor.js wordpress-static/wp-includes/js/dist/ && \
-    sed -E 's#\(\s*"iframe",#(__playground_ControlledIframe,#g' -i wordpress-static/wp-includes/js/dist/block-editor.min.js && \
-    cat /root/controlled-iframe.js wordpress-static/wp-includes/js/dist/block-editor.min.js >> /root/block-editor.min.js && \
-    mv /root/block-editor.min.js wordpress-static/wp-includes/js/dist/;
-
 # Move the static files to the final output directory
 RUN mkdir /root/output/$OUT_FILENAME
 RUN mv wordpress-static/* /root/output/$OUT_FILENAME/
